requirements/requirements.txt
attrs==22.1.0
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
bcrypt==4.0.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
et-xmlfile==1.1.0
flasgger==0.9.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.3
imageio==2.21.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
mistune==2.0.4
MyQR==2.3.1
numpy==1.23.3
openpyxl==3.0.10
pandas==1.4.4
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.41
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.8.1
